+++ src/nl/tdegroot/games/nemesis/level/Level.java
@@ -234,7 +234,7 @@ public class Level implements TileBasedMap {
 			}
 
 			if (mobCash != "") {
-				spawner.setMobCash(Integer.parseInt(mobCash));
+				spawner.setMobCash(Double.parseDouble(mobCash));
 			}
 
 		}
